pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SquareWars"
version = "1.0.0"
authors = [
    { name = "Matiiss", email = "matiiss@matiiss.com" },
    { name = "JiffyRob", email = "jiffyrob@jiffyrob.com" },
]
description = ""
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["pygame-ce==2.4.1"]

[project.urls]
Homepage = "https://github.com/Matiiss/SquareWars"
Issues = "https://github.com/Matiiss/SquareWars/issues"

[project.optional-dependencies]
dev = ["ruff"]

[project.gui-scripts]
square-wars = "square_wars:run"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "C901"]
